# docs/swagger.yaml
basePath: /api/v7
consumes:
- application/json
definitions:
  models.Bid:
    properties:
      amount:
        type: number
      user:
        $ref: '#/definitions/models.ShortUserData'
    type: object
...
  models.Listing:
    properties:
      active:
        type: boolean
      auctioneer:
        $ref: '#/definitions/models.ShortUserData'
      bids_count:
        type: integer
      category:
        type: string
      closing_date:
        type: string
      desc:
        type: string
      highest_bid:
        type: number
      image:
        type: string
      name:
        type: string
      price:
        type: number
      slug:
        type: string
      time_left_seconds:
        type: integer
      watchlist:
        type: boolean
    type: object
...
  models.ShortUserData:
    properties:
      avatar:
        example: https://my-avatar.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
...
  schemas.BidResponseDataSchema:
    properties:
      bids:
        items:
          $ref: '#/definitions/models.Bid'
        type: array
      listing:
        type: string
    type: object
...
  schemas.BidsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.BidResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.CreateListingResponseDataSchema:
    properties:
      active:
        type: boolean
      auctioneer:
        $ref: '#/definitions/models.ShortUserData'
      bids_count:
        type: integer
      category:
        type: string
      closing_date:
        type: string
      desc:
        type: string
      file_upload_data:
        $ref: '#/definitions/utils.SignatureFormat'
      highest_bid:
        type: number
      image:
        type: string
      name:
        type: string
      price:
        type: number
      slug:
        type: string
      time_left_seconds:
        type: integer
      watchlist:
        type: boolean
    type: object
  schemas.CreateListingResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.CreateListingResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.CreateListingSchema:
    properties:
      category:
        example: category_slug
        type: string
      closing_date:
        example: "2006-01-02T15:04:05.000Z"
        type: string
      desc:
        example: Product description
        type: string
      file_type:
        example: image/jpeg
        type: string
      name:
        example: Product name
        maxLength: 70
        type: string
      price:
        example: 1000
        type: number
    required:
    - category
    - closing_date
    - desc
    - file_type
    - name
    - price
    type: object
...
  schemas.ListingsResponseSchema:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Listing'
        type: array
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.ProfileResponseDataSchema:
    properties:
      avatar:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
  schemas.ProfileResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.ProfileResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.UpdateListingSchema:
    properties:
      active:
        example: true
        type: boolean
      category:
        example: category_slug
        type: string
      closing_date:
        example: "2006-01-02T15:04:05.000Z"
        type: string
      desc:
        example: Product description
        type: string
      file_type:
        example: image/jpeg
        type: string
      name:
        example: Product name
        maxLength: 70
        type: string
      price:
        example: 1000
        type: number
    type: object
  schemas.UpdateProfileResponseDataSchema:
    properties:
      file_upload_data:
        $ref: '#/definitions/utils.SignatureFormat'
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
    type: object
  schemas.UpdateProfileResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.UpdateProfileResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.UpdateProfileSchema:
    properties:
      file_type:
        example: image/png
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
    required:
    - first_name
    - last_name
    type: object
...
  utils.ErrorResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      status:
        type: string
    type: object
  utils.SignatureFormat:
    properties:
      public_id:
        example: images/f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      signature:
        example: e1ba4683fbbf90b75ca22e9f8e545b18c6b24eae
        type: string
      timestamp:
        example: 1678828200
        type: integer
    type: object
info:
  contact: {}
  description: A simple bidding API built with Fiber
  title: Bidout Auction API
  version: "7.0"
paths:
  /auctioneer:
    get:
      description: This endpoint gets the current user's profile.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProfileResponseSchema'
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - Auctioneer
    put:
      description: 'This endpoint updates an authenticated user''s profile. Note:
        use the returned upload_url to upload avatar to cloudinary'
      parameters:
      - description: Update User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateProfileSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateProfileResponseSchema'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - Auctioneer
  /auctioneer/listings:
    get:
      description: This endpoint retrieves all listings by the current user.
      parameters:
      - description: Listings Quantity
        in: query
        name: quantity
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ListingsResponseSchema'
      security:
      - BearerAuth: []
      summary: Retrieve all listings by the current user
      tags:
      - Auctioneer
    post:
      description: 'This endpoint creates a new listing. Note: Use the returned upload_url
        to upload image to cloudinary'
      parameters:
      - description: Create Listing
        in: body
        name: listing
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateListingSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CreateListingResponseSchema'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a listing
      tags:
      - Auctioneer
  /auctioneer/listings/{slug}:
    patch:
      description: 'This endpoint updates a particular listing. Note: Use the returned
        upload_url to upload image to cloudinary'
      parameters:
      - description: Listing Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Update Listing
        in: body
        name: listing
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateListingSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CreateListingResponseSchema'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a listing
      tags:
      - Auctioneer
  /auctioneer/listings/{slug}/bids:
    get:
      description: This endpoint retrieves all bids in a particular listing by the
        current user.
      parameters:
      - description: Listing Slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.BidsResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve bids in a listing (current user)
      tags:
      - Auctioneer
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer jwt_string' to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
  GuestUserAuth:
    description: For guest watchlists. Get ID (uuid) from '/api/v7/listings/watchlist'
      POST endpoint
    in: header
    name: GuestUserId
    type: apiKey
swagger: "2.0"